backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "django==5.2.1",
    "sqlparse==0.5.3",
    "djangorestframework==3.16.0",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "fonttools==4.58.1",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.3",
    "numpy==2.2.6",
    "packaging==25.0",
    "pillow==11.2.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "brotli==1.1.0",
    "cffi==1.17.1",
    "cssselect2==0.8.0",
    "pycparser==2.22",
    "pydyf==0.11.0",
    "pyphen==0.17.2",
    "tinycss2==1.4.0",
    "tinyhtml5==2.0.0",
    "weasyprint==65.1",
    "webencodings==0.5.1",
    "zopfli==0.2.3.post1",
    "amqp==5.3.1",
    "billiard==4.2.1",
    "celery==5.5.3",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "kombu==5.5.4",
    "prompt-toolkit==3.0.51",
    "redis==6.2.0",
    "tzdata==2025.2",
    "vine==5.1.0",
    "wcwidth==0.2.13"
]